Ghana Clean Team Survey

The Clean Team Ghana dataset captures extensive information on various aspects of container-based sanitation in Ghana. It includes data on household demographics, sanitation facilities, cleanliness perceptions, usage patterns, satisfaction levels, concerns, and factors influencing the adoption of Clean Team Ghana's container-based sanitation services. This dataset aims to evaluate the effectiveness, acceptance, and challenges associated with container-based sanitation practices in Ghanaian communities.

Usage

cleanteamghana

Format

A tibble with 583 rows and 127 variables:

uuid

Unique identifier for each record

submission_time

Timestamp of the survey submission

area

Area of the observation

district

District of the observation

community

Community of the observation

cleanliness_of_neighborhood

Perceived cleanliness of the neighborhood

safety_of_neighborhood

Perceived safety of the neighborhood

development_priority_of_neighborhood

Development priority of the neighborhood

age_of_respondent

Age of the respondent

sex_respondent

Gender of the respondent

marital_status

Marital status of the respondent

highest_level_of_education

Highest level of education completed by the respondent

employment_type

Type of employment of the respondent

house_ownership

Ownership status of the house

landlord_residence

Whether the landlord resides in the same location

female_less_5

Number of females under 5 years old living in the house

female_from_5_to_18

Number of females between 5 and 18 years old living in the house

female_more_18

Number of females over 18 years old living in the house

male_less_5

Number of males under 5 years old living in the house

male_from_5_to_18

Number of males between 5 and 18 years old living in the house

male_more_18

Number of males over 18 years old living in the house

number_households_sharing

Number of households sharing a facility

number_of_households

Total number of households in the house/area?

number_of_rooms

Total number of rooms in the house

daily_toilet_usage

Frequency of toilet usage per day

electricity_situation

Electricity situation in the house

fuel_used_for_cooking

Type of fuel used for cooking

location_of_water_source

Location of water source

able_to_observe_toilet

Ability to observe a toilet in the area by the interviewer

cant_afford_toilet

Inability to afford a toilet by inhabitants of the house

functioning_toilet

Presence of a functioning toilet

functioning_toilet_is_ctg

The functioning toilet is from Clean Team Ghana?

handwashing_facility

Availability of a handwashing facility

floor_material

Material used for flooring

roof_material

Material used for roofing

wall_material

Material used for walls

pan_material

Material used for toilet pans

pit_lining

Presence of pit lining

mechanical_emptying

Mechanical emptying of the pit

primary_solidwaste_collection

Primary method of solid waste collection

presence_of_drain

Presence of a drain in the toilet

presence_of_paved_road

Presence of paved roads near the toilet

presence_of_street_lights

Presence of street lights near the toilet

cleaned_toilet

Cleanliness of the toilel

cleanliness_of_pan

Cleanliness of the toilet pan

cleanliness_of_slab

Cleanliness of the toilet slab

concerns_about_using_ctg_toilet

Whether individuals have concerns about using the Clean Team Ghana (CTG) toilet

concerns_about_using_ctg_toilet_specify

Respondents can specify their concerns about using the CTG toilet if they have any

container_replaced_freq_ctg

Frequency of container replacement for the CTG toilet

cost_per_use

Cost incurred per use of the CTG toilet

ctg_cleanliness

Perceived cleanliness of the CTG toilet

ctg_information_source

Source from which individuals obtained information about the CTG toilet

ctg_issues

Any issues or problems related to the CTG toilet

ctg_recommendation

Whether individuals would recommend the CTG toilet to others

door_lockable_inside

Whether the CTG toilet has a lockable door from the inside

door_lockable_outside

Whether the CTG toilet has a lockable door from the outside

ever_unable_to_use_ctg

Whether individuals have ever been unable to use the CTG toilet

ever_unable_to_use_ctg_specify

Respondents can specify reasons if they have ever been unable to use the CTG toilet

fear_of_being_removed

Whether individuals fear the CTG toilet of being removed

how_likely_to_recommend_clean_team

Likelihood of individuals recommending Clean Team Ghana's services to others

is_ctg_available

Availability of the CTG toilet in this particular area

knowledge_of_toilet_sharing

Individuals' awareness or understanding of sharing toilets among multiple users

management_of_children_feces

Need for managing children's feces

odour_control_agents_baseline

Agents or methods used for controlling odors at the baseline or initial stage

odour_control_agents_endline

Agents or methods used for controlling odors at the endpoint or final stage

odour_control_baseline

Effectiveness or measures taken to control odors at the baseline or initial stage

odour_control_endline

Effectiveness or measures taken to control odors at the endpoint or final stage

people_using_ctg

Number of people using Clean Team Ghana's services or toilets in this particular area

rent_paid

Amount of rent paid by CTG?

sanqol_assault

Perceptions or experiences related to assault or safety concerns at the sanitation facility

sanqol_cleanliness

Perceptions or evaluations of the cleanliness of the sanitation facility.

sanqol_convenience

Level of convenience provided by the sanitation facility

sanqol_cost

Perceptions or satisfaction related to the cost of using the sanitation facility

sanqol_disgust

Feelings of disgust or repulsion associated with the sanitation facility

sanqol_health

Perceptions or concerns about health-related aspects associated with the sanitation facility

sanqol_privacy

Perceptions or satisfaction levels regarding privacy at the sanitation facility

sanqol_proud

Feelings of pride or satisfaction associated with using the sanitation facility

sanqol_safety

Perceptions or feelings of safety related to the sanitation facility

sanqol_shame

Feelings of shame associated with using the sanitation facility

sanqol_smell

Perceptions or experiences related to unpleasant smells at the sanitation facility

satisfaction_avoiding_unpleasant_odour

Satisfaction levels regarding avoiding unpleasant odors while using the facility

satisfaction_avoiding_unpleasant_odour_not_in_use

Satisfaction levels regarding avoiding unpleasant odors when the facility is not in use

satisfaction_combine_duities_and_using_toilet

Satisfaction levels while combining duties and using the toilet

satisfaction_manage_menstrual_hygiene

Satisfaction levels regarding the management of menstrual hygiene at the facility

satisfaction_towards_cost_of_latrine_usage

Satisfaction levels related to the cost of using the latrine or sanitation facility

satisfaction_towards_ease_to_access

Satisfaction levels related to the ease of access to the sanitation facility

satisfaction_towards_ease_to_use

Satisfaction levels related to the ease of using the sanitation facility

satisfaction_towards_privacy_of_toilet_use

Satisfaction levels related to the privacy offered during toilet use

satisfaction_towards_toilet_cleanliness

Satisfaction levels related to the cleanliness of the toilet

satisfaction_towards_toilet_uncomfortabity

Discomfort experienced while using the toilet

sharing_of_toilet_baseline

Sharing pattern or practices related to using the toilet at the baseline or initial stage

sharing_of_toilet_endline

Sharing pattern or practices related to using the toilet at the endpoint or final stage

slab_material

Material used for constructing the slab of the sanitation facility

slab_material_other

Additional information on the material used for constructing the slab if it's specified as 'other'

smell_reduction_system

Method implemented within the toilet facility to reduce or eliminate unpleasant odors

soap_for_handwashing

Availability of soap provided specifically for handwashing within the toilet facility

source_of_water_for_drinking

Available water source within the toilet facility intended for drinking purposes

source_of_water_for_other_purposes

Water source within the toilet facility used for non-drinking purposes such as cleaning, flushing, or other sanitary needs

time_to_access_toilet

Time taken to reach the toilet from a specific point, likely measured in minutes.

time_to_get_water

Time required to access water, likely measured in minutes

toilet_door

Presence of toilet door in the toilet facility

toilet_location_baseline

Initial or baseline location of the toilet facility

toilet_location_ctg

Location of the Clean Team Ghana toilet facility

toilet_on_compound_cant_use

Presence of a toilet within a compound that is unusable for some reason

toilet_public_usage_baseline

Usage pattern or availability of the toilet facility at the baseline

toilet_public_usage_endline

Usage pattern or availability of the toilet facility at the endpoint or final stage

toilet_roof

Presence of toilet roof in the toilet facility

toilet_scrubbedmopped_frequency

Frequency of cleaning activities like scrubbing and mopping of the toilet

toilet_use_payment_option

Payment for using the toilet

toilet_walls

Presence of toilet wall in the toilet facility

type_of_toilet

Type of toilet facility (e.g., pit latrine, flush toilet)

urinate_or_defecate_at_night

Behavior related to urination or defecation during the night

urinate_or_defecate_at_night_use_ctg

Nighttime urination or defecation patterns in the Clean Team Ghana toilet

use_non_ctg_toilet

Usage of a non-Clean Team Ghana provided toilet facility

wait_for_replacement_ctg

If the residents waiting for a replacement of the Clean Team Ghana toilet

water_for_handwashing

Availability of water for handwashing

why_not_observed

Reasons why CTG toilet wasn't observed

why_not_use_a_toilet

Reasons for not using a toilet in general

why_not_use_household_toilets

Reasons for not using a household toilet

why_not_use_household_toilets_other

Additional reasons for not using household toilets if specified as 'other'

years_living_in_current_house

Duration of time spent living in the current household

household_size

Total number of individuals living in a household

adults

Count of adults in the household

kids

Count of kids in the household

sanqol_rawindex

Raw numerical index representing the overall satisfaction in sanitation quality of life

sanqol_rawscore

Raw numerical score representing the overall satisfaction in sanitation quality of life

cost_per_month_public_baseline

Cost per month related to public toilet usage at the baseline stage